setup independent tests for Theseus
