+++ .github/workflows/CI.yml
@@ -74,7 +74,8 @@ jobs:
           arch: ${{ matrix.arch }}
       - uses: julia-actions/cache@v2
       - if: ${{ matrix.version == '1.10' }}
-        run: julia --project="libs/Theseus" -e 'using Pkg; Pkg.develop([PackageSpec(path=".")])'
+        run: |
+          julia --project="libs/Theseus" -e 'using Pkg; Pkg.develop([PackageSpec(path=".")])'
       - uses: julia-actions/julia-buildpkg@v1
         with:
           project: "libs/Theseus"
