+++ libs/Theseus/test/Project.toml
@@ -1,8 +1,14 @@
 [deps]
+Ariadne = "0be81120-40bf-4f8b-adf0-26103efb66f1"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+ParallelTestRunner = "d3525ed8-44d0-4b2c-a655-542cee43accc"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Theseus = "7f538e44-2768-4ef2-af90-2110c0378286"
 
 [sources]
+Ariadne = {path = "../../.."}
 Theseus = {path = ".."}
+
+[compat]
+ParallelTestRunner = "2"
